Move MetroCard to components/index/cards
