+++ pages/cards/_card.vue
@@ -59,7 +59,7 @@ import InspectionPersonsNumberCard from '@/components/index/cards/InspectionPers
 import TestedNumberCard from '@/components/index/cards/TestedNumberCard.vue'
 import TelephoneAdvisoryReportsNumberCard from '@/components/index/cards/TelephoneAdvisoryReportsNumberCard.vue'
 import ConsultationDeskReportsNumberCard from '@/components/index/cards/ConsultationDeskReportsNumberCard.vue'
-import MetroCard from '@/components/cards/MetroCard.vue'
+import MetroCard from '@/components/index/cards/MetroCard.vue'
 import AgencyCard from '@/components/cards/AgencyCard.vue'
 
 export default {
